Initial Completion of Sprint Tasks for Updating User Information
Reused userforms.html for handling the two different forms adding user or editing user, currently authtype and accounttype do not have current values as selecting the correct current values is conceptually difficult to accomplish. Addition of function in handler and views to handle the back and front end tasks of the new endpoint.